+++ CHANGELOG.rst
@@ -15,6 +15,10 @@ odd bug fix may not get mentioned here at all.
     unhandled by wed would still be handled as an editing operation by the
     browser.
 
+  - When pasting XML, wed no longer checks whether the added XML causes
+    errors. The check prevented some legitimate use-case scenarios. The user can
+    fix the errors after the fact.
+
 * 2.0.1:
 
   - Fix validation restarts. There was a bug whereby wed would not restart